UPSTREAM: mb/google/volteer: Set up SATAXPCIE1 IOSSTATE based on detected device
There is an issue with the storage device being mis-detected on exit
from S0ix which is causing the root device to disappear if the power
is actually turned off via RTD3.

To work around this read the RX state of the pin and apply the IOSSTATE
setting to drive a 0 or 1 back to the internal controller.  This will
ensure the device is detected the same on resume as on initial boot.

BUG=b:171993054
TEST=boot on volteer with PCIe NVMe and SATA SSD installed in the M.2
slot and ensure this pin is configured appropriately.  Additionally
test with PCIe RTD3 enabled to ensure suspend/resume works reliably.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 98a9f1f61de45fdf936d51342c6a2f06e1980c58
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I85542151eebd0ca411e2c70d8267a8498becee78
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47255
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Shreesh Chhabbi <shreesh.chhabbi@intel.corp-partner.google.com>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: If3694887d0b17bcebe83fa9993838aab415a3142
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2553593
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>