UPSTREAM: libpayload/i8042: Increase response timeout to 1.5s
The current timeout of 500ms is too low. For instance self-test
of the KBC integrated into IT8516E took almost 1s in tests. We
already check for presence of the KBC before the self-test. So
the timeout should only trigger on a hardware defect and we can
leave some margin.

BUG=none
BRANCH=none
TEST=none

Change-Id: Idfcefec27bbd361cb9ac1bba268a18d840b47564
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0748a87455eed53fee0858ccc74291e8e2357cd0
Original-Change-Id: I95f01a4e605a9c7deb894a71e102c3a881759bb1
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47588
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2540327
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>