UPSTREAM: mb/supermicro/x11ssh-tf: configure "POST complete" gpio for IPMI driver
Configure the "POST complete" gpio in the devicetree for the BMC/IPMI
driver and set the pad's initial value to 0 since the signal is active-
high and shall be set by the IPMI/BMC driver.

Also add the pad to early gpio config, since it is expected to have an
external pull-up like X11SSM-F, which is wrong and would confuse the BMC.

Test: Boot the machine via the BMC web interface and check that sensors
get read correctly by the IPMI firmware when the payload starts.

Tested successfully.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 6c3ba50a444d969c3086ab18d5123f119295748b
Original-Change-Id: If344b2271bfc8d50b8b64847109818f96f2abbcb
Original-Tested-by: Patrick Rudolph <siro@das-labor.org>
Original-Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48711
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I5a4aa2f7c5f458abc3a999d7c2682a8812ac55da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2607519
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>