UPSTREAM: soc/intel/common: Add Kconfig to enable the CSE FW Update feature
Add the Kconfig to enable the CSE FW Update feature and also to
ensure all the configs are set by the mainboards to enable this
feature.
This config by default disables the CSE FW update feature for JSL
and TGL platforms. It will be enabled after splitting and including
the CSE RW and CSE RW metadata blobs in the CBFS.

BUG=b:169077783

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: f99055266bf2dba0cd85ade3bb0ba3ddeec5f7c7
Original-Change-Id: I12810031224f79aba8a4057725ae0ed5a9b36d7e
Original-Signed-off-by: V Sowmya <v.sowmya@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47523
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Change-Id: I68fc787cf86bdcc0cdcfad1e7ca507c8eae4251b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2546050
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>