UPSTREAM: mb/google/zork: Remove 50ms WIFI delay
As a part of trying to get our boot time as low as possible, any delays
in the code should try to be refactored out.  This removes the 50ms
delay in the WIFI sequence by enabling power and putting the wifi module
into reset in bootblock, then bringing it out of reset in ramstage.
This is significantly longer than the 50ms requirement.  The reset GPIO
was already being set high in ramstage, so that code didn't need to be
added.

BUG=b:171513520
TEST=Boot on boards with different module types, WIFI works on both.
BRANCH=Zork

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: cdd7d18120a1a76df4a9f8d78391c7018013d916
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I211d3da338ad368d1f011f03cf7d05121c057075
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47719
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Change-Id: I8bced12f85183dcc66066523af078646a09a1034
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2554048
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>