UPSTREAM: cbfstool: Move alignment/baseaddress handling into cbfs_add_component()
The --alignment flag is currently only handled by cbfstool add, but
there seems little reason to not handle it for all file-adding commands
(the help text actually mentions it for add-stage as well but it doesn't
currently work there). This patch moves the related code (and the
related baseaddress handling) into cbfs_add_component(). As a nice side
effect this allows us to rearrange cbfs_add_component() such that we can
conclusively determine whether we need a hash attribute before trying to
align the file, allowing that code to correctly infer the final header
size even when a hash attribute was implicitly added (for an image built
with CBFS verification enabled).

BUG=none
BRANCH=none
TEST=none

Change-Id: I4f275f7f4d2f621ced8829e8ba32bc964e76749a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c24db001efb669c06489c55122e85a4a8948b539
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Idc6d68b2c7f30e5d136433adb3aec5a87053f992
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47823
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2765534
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>