UPSTREAM: vc/amd/pi/00670F00: Add raw AGESA binary only to COREBOOT CBFS
If AGESA is added as a raw binary (and not a stage), then cbfstool
does not perform relocation. In this case, it should be added only to
COREBOOT (i.e. default) CBFS since the binary needs to be present only
in one specific location that is present in the default CBFS.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9aa52e739d91e637a7bffd5808aa57845ee06591
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6c8ba9b9ae0a0eb04f6c7be3ba8185f49db13d00
Original-Change-Id: I7a7edc217663f9d1d36b05308bbd35f56a28b9b1
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47832
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2554448
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>