UPSTREAM: soc/intel/common/block/gpio: add code for NMI enabling
Especially server boards, like the Supermicro X11SSM-F, often have a NMI
button and NMI functionality that can be triggered via IPMI. The purpose
of this is to cause the OS to create a system crashdump from a hang
system or for debugging.

Add code for enabling NMI interrupts on GPIOs configured with
PAD_CFG_GPI_NMI. The enabling mechanism is the same as SMI, so the SMI
function was copied and adapted. The `pad_community` struct gained two
variables for the registers.

Also register the NMI for LINT1 in the MADT in accordance to ACPI spec.

Test: Linux detects the NMI correctly in dmesg:
[    0.053734] ACPI: LAPIC_NMI (acpi_id[0xff] high edge lint[0x1])

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic1aecedef7f1f0a893eb231f7e9e51251280bc21
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 14512f9a9be08d945830ddb6d0e48f4d6fc1635c
Original-Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Change-Id: I4fc1a35c99c6a28b20e08a80b97bb4b8624935c9
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/48090
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2577263
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>