UPSTREAM: fw_config: Use UNDEFINED_FW_CONFIG to mean unprovisioned
A mainboard might want to configure some things differently when a
device is in an unprovisioned state. In the case when fw_config comes
from the Chromium EC, an unprovisioned device will not have a FW_CONFIG
tag in its CBI. This patch will set the fw_config value to
UNDEFINED_FW_CONFIG in the case of an error retrieving the value, as
well as adding a function, `fw_config_is_provisioned()` to indicate the
provisioning status.

BUG=none
TEST=remove fw_config from chromium EC CBI, add code to mainboard to
print return value of fw_config_is_provisioned() (`0`), add
fw_config back to CBI, run same test and see `1`.

Signed-off-by: Joe Tessler <jrt@google.com>
Original-Commit-Id: 473bc8c892c3acf6b1a222a09a53d5a190611bba
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: Ib3046233667e97a5f78961fabacbeb3099b3d442
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47956
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I39040122ba1cba323b6820789ba497a6d3777175
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2587001
Reviewed-by: Joe Tessler <jrt@chromium.org>
Tested-by: Joe Tessler <jrt@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>